backend/Source/requirements.txt
asgiref==3.3.1
certifi==2020.12.5
cffi==1.14.4
chardet==3.0.4
cryptography==3.3.1
Cython==0.29.14
Django==3.1.4
django-cors-headers==3.2.1
django-rest-knox==4.1.0
djangorestframework==3.12.2
idna==2.8
psycopg2-binary==2.8.6
pycparser==2.20
pytz==2020.5
requests==2.22.0
six==1.15.0
sqlparse==0.4.1
urllib3==1.25.11
